refactor(desktop): split ipc/database.ts into per-entity handler modules
